requirements.txt
pyperclip>=1.8.2
keyboard>=0.13.5
pillow>=9.0.0
pystray>=0.19.4
ttkthemes>=3.2.2
cryptography>=38.0.4
python-dateutil>=2.8.2
beautifulsoup4>=4.11.1
markdown>=3.4.1
python-docx>=0.8.11
pypdf>=3.0.0
rich>=13.0.0
customtkinter>=5.1.2
darkdetect>=0.8.0